+++ src/assets/ellipse.png
