+++ src/assets/image.ts
@@ -1,3 +1,3 @@
-export { default as visaCard } from "../../public/assets/visaCard.png";
-export { default as masterCard } from "../../public/assets/masterCard.png";
-export { default as ellipse } from "../../public/assets/ellipse.png";
+// export { default as visaCard } from "../../public/assets/visaCard.png";
+// export { default as masterCard } from "../../public/assets/masterCard.png";
+// export { default as ellipse } from "../../public/assets/ellipse.png";
